feat: 토스페이먼츠 결제

+++ 2Week_Mission/src/main/java/likelion/finalproject/market/member/domain/Member.java
@@ -45,7 +45,7 @@ public class Member extends BaseTimeEntity {
     private String password;
 
     @Column
-    private int cash = 0;
+    private long cash = 0;
 
     @Column
     @Enumerated(EnumType.STRING)
@@ -70,11 +70,11 @@ public class Member extends BaseTimeEntity {
         this.auth = Auth.WRITER;
     }
 
-    public void addCash(int money) {
+    public void addCash(long money) {
         this.cash += money;
     }
 
-    public void useCash(int money) {
+    public void useCash(long money) {
         this.cash -= money;
     }
 }
